fix: layout on large fonts and relocate files
